consts moved to their own file
